--- apstools/_plans/__init__.py
@@ -1,2 +1,0 @@
-from .doc_run import addDeviceDataAsStream
-from .doc_run import documentation_run
