+++ apstools/plans/doc_run.py
@@ -86,7 +86,6 @@ def documentation_run(text, stream=None, bec=None, md=None):
 
     return uid
 
-
 # -----------------------------------------------------------------------------
 # :author:    Pete R. Jemian
 # :email:     jemian@anl.gov
