+++ apstools/snapshot/application.py
@@ -49,9 +49,9 @@ import tkinter.ttk as ttk
 import databroker
 
 from . import textreadonly
-from .. import utils as APS_utils
-from .. import plans as APS_plans
-from .. import callbacks as APS_callbacks
+from .. import utils
+from .. import plans
+from .. import callbacks
 
 
 BROKER_CONFIG = "mongodb_config"
@@ -155,18 +155,18 @@ def snapshot_cli():
     md = OrderedDict(purpose="archive a set of EPICS PVs")
     md.update(parse_metadata(args))
 
-    obj_dict = APS_utils.connect_pvlist(args.EPICS_PV, wait=False)
+    obj_dict = utils.connect_pvlist(args.EPICS_PV, wait=False)
     time.sleep(2)  # FIXME: allow time to connect
 
     db = databroker.Broker.named(args.broker_config)
     RE = RunEngine({})
     RE.subscribe(db.insert)
 
-    uuid_list = RE(APS_plans.snapshot(obj_dict.values(), md=md))
+    uuid_list = RE(plans.snapshot(obj_dict.values(), md=md))
 
     if args.report:
         snap = list(db(uuid_list[0]))[0]
-        APS_callbacks.SnapshotReport().print_report(snap)
+        callbacks.SnapshotReport().print_report(snap)
 
 
 class Capturing(list):  # LGTM
