+++ docs/source/api/_plans.rst
@@ -1,16 +1,86 @@
-
+===========
 Plans
------
+===========
+
+Plans that might be useful at the APS when using Bluesky.
+
+Categories
+----------
+
+.. _plans.batch:
+
+BATCH SCANNING
++++++++++++++++++++
+
+.. autosummary::
+
+   ~apstools.plans.command_list.execute_command_list
+   ~apstools.plans.command_list.get_command_list
+   ~apstools.plans.command_list.parse_Excel_command_file
+   ~apstools.plans.command_list.parse_text_command_file
+   ~apstools.plans.command_list.register_command_handler
+   ~apstools.plans.command_list.run_command_file
+   ~apstools.plans.command_list.summarize_command_file
+
+.. _plans.custom:
+
+CUSTOM SCANS
++++++++++++++++++++
+
+.. autosummary::
+
+   ~apstools.plans.doc_run.documentation_run
+   ~apstools.plans.alignment.lineup
+   ~apstools.plans.nscan_support.nscan
+   ~apstools.plans.snapshot_support.snapshot
+   ~apstools.plans.sscan_support.sscan_1D
+   ~apstools.plans.alignment.TuneAxis
+   ~apstools.plans.alignment.tune_axes
+
+.. _plans.overall:
+
+OVERALL
++++++++++++++++++++
+
+.. autosummary::
+
+   ~apstools.plans.doc_run.addDeviceDataAsStream
+   ~apstools.plans.doc_run.documentation_run
+   ~apstools.plans.command_list.execute_command_list
+   ~apstools.plans.command_list.get_command_list
+   ~apstools.plans.alignment.lineup
+   ~apstools.plans.nscan_support.nscan
+   ~apstools.plans.command_list.parse_Excel_command_file
+   ~apstools.plans.command_list.parse_text_command_file
+   ~apstools.plans.command_list.register_command_handler
+   ~apstools.plans.command_list.run_command_file
+   ~apstools.plans.snapshot_support.snapshot
+   ~apstools.plans.sscan_support.sscan_1D
+   ~apstools.plans.command_list.summarize_command_file
+   ~apstools.plans.alignment.TuneAxis
+   ~apstools.plans.alignment.tune_axes
 
 Also consult the :ref:`Index <genindex>` under the *Bluesky* heading
 for links to the Callbacks, Devices, Exceptions, and Plans described
 here.
 
-.. automodule:: apstools.plans
+Submodules
+---------------
+
+.. automodule:: apstools.plans.alignment
     :members:
 
-Submodules
-++++++++++
+.. automodule:: apstools.plans.command_list
+    :members:
 
-.. automodule:: apstools._plans.doc_run
+.. automodule:: apstools.plans.doc_run
+    :members:
+
+.. automodule:: apstools.plans.nscan_support
+    :members:
+
+.. automodule:: apstools.plans.snapshot_support
+    :members:
+
+.. automodule:: apstools.plans.sscan_support
     :members:
